将所有id类型改为Long，实现人力资源档案管理中登记、删除、恢复已删除、变更、查询

+++ rear/src/main/java/com/hrsys/controller/salary/SalaryDetailController.java
@@ -42,7 +42,7 @@ public class SalaryDetailController {
 
     @DeleteMapping("/{id}")
     @ApiOperation(value = "删除薪酬发放明细")
-    public Result<?> delete(@PathVariable @NotNull(message = "id不能为空") Integer id) {
+    public Result<?> delete(@PathVariable @NotNull(message = "id不能为空") Long id) {
         hrRecordService.removeById(id);
         return Result.success();
     }
